feat(player): 플레이어 hp, currentHp 변수 int형에서 float형으로 수정
이후 Lerp 함수를 사용해 HP 바 감소를 원활하게 구현하기 위해서 float형으로 변경함.

+++ Assets/Scripts/PlayerController.cs
@@ -15,8 +15,8 @@ public class PlayerController : MonoBehaviour
 
     //플레이어의 체력
     [SerializeField]
-    private int hp;
-    private int currentHp;
+    private float hp;
+    private float currentHp;
 
     // 플레이어의 스테미나
     [SerializeField]
@@ -155,12 +155,12 @@ public class PlayerController : MonoBehaviour
 
     //플레이어 정보 내보내는 메서드들
     #region GetMethods
-    public int GetPlayerCurrentHP()
+    public float GetPlayerCurrentHP()
     {
         return currentHp;
     }
 
-    public int GetPlayerHP()
+    public float GetPlayerHP()
     {
         return hp;
     }
